fix: read --health-check-cache-ttl through the accessor that matches its flag type
urfave/cli v3 type-asserts strictly, so Int64() on an IntFlag returns 0.
The guard read Int() and the assignment read Int64(), so passing the flag
set CacheTTL to 0 and discarded whatever the yaml had; Prepare then
rescued it to 300 and the flag never took effect.

+++ agent.go
@@ -195,7 +195,7 @@ func main() {
 				Usage:   "timeout for agent to check container's health status",
 				Sources: cli.EnvVars("ERU_AGENT_HEALTH_CHECK_TIMEOUT"),
 			},
-			&cli.IntFlag{
+			&cli.Int64Flag{
 				Name:    "health-check-cache-ttl",
 				Usage:   "ttl for container's health status in local memory",
 				Sources: cli.EnvVars("ERU_AGENT_HEALTH_CHECK_CACHE_TTL"),
